Let the createReducer worry about current state
